Implementing lazymap for Seq
Finally, using BufferedSeq as the baseline, I have a plausible way
to implement lazy bulk operations on Seqs, using the base abstract
template of BaseLazySeq. It should be possible to add all bulk
operations directly into the base Seq interface and hide the use
of InfiniteSeq and LazySeq for static factories over Iterators and
Suppliers, which should allow to only keep the base abstract class
public.