Adding Binary tree heap implementation
(work-in-progress). About the interface, how should I related Heap
Col and Seq together. Is it useful to have the Heap interface be
F-bounded to return the same Heap implem after calling pop() ? If so,
then Heap cannot be declared as implementing Col because union would
then be for all different Heap implementations, which would be
inconvenient. For the user, it is probably better for Heap not to be
F bounded. In this case, what to do with pop() ? Could Heap be
a special case of Seq ?